test(wallet-core): refactor tests, move rbf fee formatting test
